+++ Projects/QuadraticElements/QuadraticGeometricTests.cpp
@@ -315,8 +315,8 @@ int main(int argc, char * const argv[])
     }
     
     //inserting quadrilaterals
-    int elId = 0;
-    TPZVec <int> Topol(4);
+    long elId = 0;
+    TPZVec <long> Topol(4);
     
     for(int n = 0; n < 4; n++) Topol[n] = n;
     new TPZGeoElRefPattern< pzgeom::TPZGeoQuad > (elId,Topol,matElId,*gmesh);
